// frontend/src/lib/api/model/paperBlock.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * e-proshno API
 * OpenAPI spec version: v1
 */
import type { PaperBlockKind } from './paperBlockKind';
import type { PaperQuestion } from './paperQuestion';

export interface PaperBlock {
  kind: PaperBlockKind;
  /** @nullable */
  stimulus?: string | null;
  questions: PaperQuestion[];
}