// frontend/src/lib/api/model/paperQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * e-proshno API
 * OpenAPI spec version: v1
 */
import type { McqKind } from './mcqKind';
import type { PaperCqPart } from './paperCqPart';
import type { PaperOption } from './paperOption';
import type { QuestionType } from './questionType';

export interface PaperQuestion {
  number: number;
  questionId: string;
  type: QuestionType;
  mcqKind?: McqKind | null;
  stem: string;
  options: PaperOption[];
  cqParts: PaperCqPart[];
  /** @nullable */
  explanation?: string | null;
  boardTags: string[];
  importance: number;
  marks: number;
}